requirements.txt
python-binance
ta
pandas
requests
aiohttp
backoff
python-dotenv